DF-15194 Some fixes.
Former-commit-id: b0b6a2fb2c8c3ef25ca1d776cea6ae82f28c185c

+++ Sources/Internal/Scene3D/Systems/ParticleEffectSystem.h
@@ -20,7 +20,7 @@ public:
         Texture* noise = nullptr;
         Texture* alphaRemapTexture = nullptr;
         eBlending blending = BLENDING_ALPHABLEND;
-        uint32 layerId = 1;
+        uint64 layerId = 1;
         bool enableFog = false;
         bool enableFrameBlend = false;
         bool enableFlow = false;
